+++ app/code/community/Nosto/Tagging/Model/Meta/Product.php
@@ -36,7 +36,6 @@
  */
 class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
 {
-
     use Nosto_Tagging_Model_Meta_Product_Trait;
 
     /**
@@ -85,6 +84,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
      *
      * @param Mage_Catalog_Model_Product $product the product model.
      * @param Mage_Core_Model_Store|null $store the store to get the product data for.
+     * @return bool
      * @throws Nosto_NostoException
      */
     public function loadData(Mage_Catalog_Model_Product $product, Mage_Core_Model_Store $store = null)
@@ -136,6 +136,9 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
         } elseif ($dataHelper->isVariationEnabled($store)) {
             $this->amendVariations($product, $store);
         }
+        if ($dataHelper->getUseCustomFields($store)) {
+            $this->setCustomFields($this->loadCustomFields($product));
+        }
 
         Mage::dispatchEvent(
             Nosto_Tagging_Helper_Event::EVENT_NOSTO_PRODUCT_LOAD_AFTER,
@@ -144,6 +147,8 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
                 'magentoProduct' => $product
             )
         );
+
+        return true;
     }
 
     /**
@@ -152,12 +157,13 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
      *
      * @param Mage_Catalog_Model_Product $product
      * @param Mage_Core_Model_Store $store
+     * @throws Nosto_NostoException
      */
     protected function amendVariations(Mage_Catalog_Model_Product $product, Mage_Core_Model_Store $store)
     {
-        /* @var Nosto_Tagging_Helper_Variation $variationHelper  */
+        /* @var Nosto_Tagging_Helper_Variation $variationHelper */
         $variationHelper = Mage::helper('nosto_tagging/variation');
-        $this->setVariationId($variationHelper->getDefaultVariationId($store));
+        $this->setVariationId($variationHelper->getDefaultVariationId());
 
         /** @var Nosto_Tagging_Model_Meta_Variation_Collection $variationCollection */
         $variationCollection = Mage::getModel('nosto_tagging/meta_variation_collection');
@@ -308,7 +314,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
      */
     protected function amendReviews(Mage_Catalog_Model_Product $product, Mage_Core_Model_Store $store)
     {
-        /* @var Nosto_Tagging_Helper_Data $dataHelper*/
+        /* @var Nosto_Tagging_Helper_Data $dataHelper */
         $dataHelper = Mage::helper('nosto_tagging');
         $ratingProvider = $dataHelper->getRatingsAndReviewsProvider($store);
         if ($ratingProvider) {
@@ -339,6 +345,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
      *
      * @param Mage_Catalog_Model_Product $product the product model.
      * @param Mage_Core_Model_Store $store the store model.
+     * @throws Nosto_NostoException
      */
     protected function amendAttributeTags(Mage_Catalog_Model_Product $product, Mage_Core_Model_Store $store)
     {
@@ -433,7 +440,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
         if (method_exists($this, $compatibilityMethod)) {
             return $this->$compatibilityMethod($args);
         } else {
-            trigger_error('Call to undefined method '.__CLASS__.'::'.$method.'()', E_USER_ERROR); // @codingStandardsIgnoreLine
+            trigger_error('Call to undefined method ' . __CLASS__ . '::' . $method . '()', E_USER_ERROR); // @codingStandardsIgnoreLine
         }
     }
 
@@ -620,9 +627,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
         $reloadedProduct = $productModel->setStoreId($store->getId())->load($product->getId());
         if ($reloadedProduct instanceof Mage_Catalog_Model_Product) {
             try {
-                $this->loadData($reloadedProduct, $store);
-
-                return true;
+                return $this->loadData($reloadedProduct, $store);
             } catch (Nosto_NostoException $e) {
                 Nosto_Tagging_Helper_Log::exception($e);
             }